requirements.txt
django
django-mptt>=0.9
djangorestframework
requests
swapper
